Issue #335: prevent code duplication through the introduction of a base class IModelItem used as a base by ITagModelItem and INotebookModelItem
