+++ lib/model/notebook/INotebookModelItem.cpp
@@ -31,82 +31,6 @@
 
 namespace quentier {
 
-void INotebookModelItem::setParent(INotebookModelItem * parent)
-{
-    m_pParent = parent;
-
-    if (m_pParent->rowForChild(this) < 0) {
-        m_pParent->addChild(this);
-    }
-}
-
-INotebookModelItem * INotebookModelItem::childAtRow(const int row) const
-{
-    if ((row < 0) || (row >= m_children.size())) {
-        return nullptr;
-    }
-
-    return m_children[row];
-}
-
-int INotebookModelItem::rowForChild(const INotebookModelItem * pChild) const
-{
-    return m_children.indexOf(const_cast<INotebookModelItem *>(pChild));
-}
-
-void INotebookModelItem::insertChild(const int row, INotebookModelItem * pItem)
-{
-    if (Q_UNLIKELY(!pItem)) {
-        return;
-    }
-
-    pItem->m_pParent = this;
-    m_children.insert(row, pItem);
-}
-
-void INotebookModelItem::addChild(INotebookModelItem * pItem)
-{
-    if (Q_UNLIKELY(!pItem)) {
-        return;
-    }
-
-    pItem->m_pParent = this;
-    m_children.push_back(pItem);
-}
-
-bool INotebookModelItem::swapChildren(const int srcRow, const int dstRow)
-{
-    if ((srcRow < 0) || (srcRow >= m_children.size()) || (dstRow < 0) ||
-        (dstRow >= m_children.size()))
-    {
-        return false;
-    }
-
-#if QT_VERSION >= QT_VERSION_CHECK(5, 13, 0)
-    m_children.swapItemsAt(srcRow, dstRow);
-#else
-    m_children.swap(srcRow, dstRow);
-#endif
-
-    return true;
-}
-
-INotebookModelItem * INotebookModelItem::takeChild(const int row)
-{
-    if ((row < 0) || (row >= m_children.size())) {
-        return nullptr;
-    }
-
-    INotebookModelItem * pItem = m_children.takeAt(row);
-    if (pItem) {
-        pItem->m_pParent = nullptr;
-    }
-
-    return pItem;
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
 QDebug & operator<<(QDebug & dbg, const INotebookModelItem::Type type)
 {
     using Type = INotebookModelItem::Type;
