+++ lib/model/tag/ITagModelItem.cpp
@@ -27,82 +27,6 @@
 
 namespace quentier {
 
-void ITagModelItem::setParent(ITagModelItem * pParent)
-{
-    m_pParent = pParent;
-
-    if (m_pParent->rowForChild(this) < 0) {
-        m_pParent->addChild(this);
-    }
-}
-
-ITagModelItem * ITagModelItem::childAtRow(const int row) const
-{
-    if ((row < 0) || (row >= m_children.size())) {
-        return nullptr;
-    }
-
-    return m_children[row];
-}
-
-int ITagModelItem::rowForChild(const ITagModelItem * pChild) const
-{
-    return m_children.indexOf(const_cast<ITagModelItem *>(pChild));
-}
-
-void ITagModelItem::insertChild(const int row, ITagModelItem * pItem)
-{
-    if (Q_UNLIKELY(!pItem)) {
-        return;
-    }
-
-    pItem->m_pParent = this;
-    m_children.insert(row, pItem);
-}
-
-void ITagModelItem::addChild(ITagModelItem * pItem)
-{
-    if (Q_UNLIKELY(!pItem)) {
-        return;
-    }
-
-    pItem->m_pParent = this;
-    m_children.push_back(pItem);
-}
-
-bool ITagModelItem::swapChildren(const int sourceRow, const int destRow)
-{
-    if ((sourceRow < 0) || (sourceRow >= m_children.size()) || (destRow < 0) ||
-        (destRow >= m_children.size()))
-    {
-        return false;
-    }
-
-#if QT_VERSION >= QT_VERSION_CHECK(5, 13, 0)
-    m_children.swapItemsAt(sourceRow, destRow);
-#else
-    m_children.swap(sourceRow, destRow);
-#endif
-
-    return true;
-}
-
-ITagModelItem * ITagModelItem::takeChild(const int row)
-{
-    if ((row < 0) || (row >= m_children.size())) {
-        return nullptr;
-    }
-
-    auto * pItem = m_children.takeAt(row);
-    if (pItem) {
-        pItem->m_pParent = nullptr;
-    }
-
-    return pItem;
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
 QDebug & operator<<(QDebug & dbg, const ITagModelItem::Type type)
 {
     using Type = ITagModelItem::Type;
